requirements.txt
gymnasium==0.29.1
numpy==1.26.4
pygame==2.5.2
pyopengl==3.1.7
imageio==2.34.0
pillow==10.2.0
stable-baselines3==2.1.0
tensorboard==2.15.1